+++ django/applications/catmaid/static/js/widgets/confirmation-3d-dialog.js
@@ -182,10 +182,10 @@
       strahlerShadingCb.setAttribute('class', 'ui-button');
       strahlerShadingCb.checked = false;
       strahlerShadingCb.onchange = function() {
-        var shading = this.checked ? 'strahler' :'active_node_split';
+        var shading = this.checked ? 'strahler' : (self.shadingMethod || 'active_node_split');
         self.webglapp.options.shading_method = shading;
-        self.webglapp.updateSkeletonColors();
-        self.webglapp.render();
+        self.webglapp.updateSkeletonColors()
+          .then(self.webglapp.render.bind(self.webglapp));
         self.onSettingChanged('strahler-shading', this.checked);
       };
       var strahlerShading = document.createElement('label');